refactor(solver): split tuple-failure explain out; narrow union-source set
Addresses review on #12290:

- Split `explain_tuple_failure`/`tuple_element_type_mismatch` out of
  `explain.rs` into a sibling `explain_tuple.rs` module. `explain.rs` was
  already over the 2000-line source-shard cap on main (2058) and the
  union-source change pushed it to 2070; the extraction brings it to 1806
  (no behavior change — pure relocation, `explain_tuple_failure` is now
  `pub(super)` and called from `explain_failure_inner` as before).

- Narrow the union-source member-elaboration allowlist to the two families
  verified to render exactly like `tsc`: `ArrayElementMismatch` and
  `ReadonlyToMutableAssignment`. Drop `OptionalPropertyRequired` and
  `ReadonlyPropertyMismatch`: the former only arises under
  `exactOptionalPropertyTypes` and `tsc` leads it with the member header
  (it needs the header-led path, not the self-heading one); the latter
  could not be reached as a first-failing union member (those shapes route
  through `MissingProperty`/`PropertyTypeMismatch`, already covered). The
  exclusion is documented in the allowlist comment.

Tests unchanged: the existing array-element and readonly-tuple cases cover
both retained families. union_source_structural 10/10, missing_property
5/5, tuple_source 8/8; ts2322 287/4 (4 pre-existing).
