+++ crates/tsz-solver/src/relations/subtype/mod.rs
@@ -14,6 +14,7 @@ pub(crate) mod cache;
 pub(crate) mod core;
 pub(crate) mod explain;
 pub(crate) mod explain_function;
+pub(crate) mod explain_tuple;
 pub(crate) mod helpers;
 pub(crate) mod overlap;
 pub(crate) mod rules;
